+++ receiver/windowseventlogreceiver/doc.go
@@ -14,4 +14,4 @@
 
 // Package stanzareceiver implements a receiver that can be used by the
 // Opentelemetry collector to receive logs using the stanza log agent
-package windowslogreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowslogreceiver"
+package windowseventlogreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowseventlogreceiver"
