+++ receiver/windowseventlogreceiver/testdata/config.yaml
