+++ receiver/windowseventlogreceiver/windows_receivertype.go
@@ -1,0 +1,36 @@
+// Copyright The OpenTelemetry Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build windows
+// +build windows
+
+package windowseventlogreceiver // import "github.com/open-telemetry/opentelemetry-collector-contrib/receiver/windowseventlogreceiver"
+
+import (
+	"github.com/open-telemetry/opentelemetry-log-collection/operator"
+	"github.com/open-telemetry/opentelemetry-log-collection/operator/input/windows"
+	"go.opentelemetry.io/collector/config"
+	"gopkg.in/yaml.v2"
+)
+
+// DecodeInputConfig unmarshals the input operator
+func (f ReceiverType) DecodeInputConfig(cfg config.Receiver) (*operator.Config, error) {
+	logConfig := cfg.(*WindowsLogConfig)
+	yamlBytes, _ := yaml.Marshal(logConfig.Input)
+	inputCfg := windows.NewDefaultConfig()
+	if err := yaml.Unmarshal(yamlBytes, &inputCfg); err != nil {
+		return nil, err
+	}
+	return &operator.Config{Builder: inputCfg}, nil
+}
