--- receiver/windowslogreceiver/testdata/simple.log
@@ -1,3 +1,0 @@
-2020-08-25 INFO Something routine
-2020-08-25 ERROR Something bad happened!
-2020-08-25 DEBUG Some details...
